requirements.txt
openai==0.28
langchain==0.1.7
psycopg2-binary
aiogram==3.0.0b7
beautifulsoup4
fake_useragent
requests
lxml